Use SierraEntryPoint in ContractClass.
