+++ crates/starknet/src/casm_contract_class.rs
@@ -1,7 +1,9 @@
 use std::collections::HashMap;
 
-use num_bigint::BigInt;
-use serde::{Deserialize, Serialize};
+use num_bigint::{BigInt, BigUint};
+use num_traits::Num;
+use serde::ser::Serializer;
+use serde::{Deserialize, Deserializer, Serialize};
 use sierra_gas::{calc_gas_info, CostError};
 use sierra_to_casm::compiler::CompilationError;
 use sierra_to_casm::metadata::Metadata;
@@ -54,5 +56,33 @@ impl CasmContractClass {
         // TODO(ilya): Fix entry points.
 
         Ok(Self { bytecode, hints, entry_points_by_type: ContractEntryPoints::default() })
+    }
+}
+
+#[derive(Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ContractEntryPoint {
+    /// A field element that encodes the signature of the called function.
+    #[serde(serialize_with = "serialize_big_uint", deserialize_with = "deserialize_big_uint")]
+    pub selector: BigUint,
+    /// The offset of the instruction that should be called within the contract bytecode.
+    pub offset: usize,
+}
+
+pub fn serialize_big_uint<S>(num: &BigUint, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&format!("{:#x}", num))
+}
+
+pub fn deserialize_big_uint<'a, D>(deserializer: D) -> Result<BigUint, D::Error>
+where
+    D: Deserializer<'a>,
+{
+    let s = &String::deserialize(deserializer)?;
+    match s.strip_prefix("0x") {
+        Some(num_no_prefix) => BigUint::from_str_radix(num_no_prefix, 16)
+            .map_err(|error| serde::de::Error::custom(format!("{}", error))),
+        None => Err(serde::de::Error::custom(format!("{s} does not start with `0x` is missing."))),
     }
 }
