+++ crates/starknet/src/contract_class_test.rs
@@ -3,14 +3,11 @@ use num_bigint::BigUint;
 use pretty_assertions::assert_eq;
 
 use crate::abi;
-use crate::contract_class::{ContractClass, ContractEntryPoint, ContractEntryPoints};
+use crate::contract_class::{ContractClass, ContractEntryPoints, SierraEntryPoint};
 
 #[test]
 fn test_serialization() {
-    let external = vec![ContractEntryPoint {
-        selector: BigUint::from(u128::MAX),
-        offset: BigUint::from(1u32),
-    }];
+    let external = vec![SierraEntryPoint { selector: BigUint::from(u128::MAX), function_id: 7 }];
 
     let contract = ContractClass {
         sierra_program: sierra::program::Program {
@@ -39,7 +36,7 @@ fn test_serialization() {
             "EXTERNAL": [
               {
                 "selector": "0xffffffffffffffffffffffffffffffff",
-                "offset": "0x1"
+                "function_id": 7
               }
             ],
             "L1_HANDLER": [],
